inference/pyproject.toml
[project]
name = "generate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers>=0.36.0",
    "ftfy>=6.3.1",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "peft>=0.18.0",
    "pillow>=12.0.0",
    "safetensors>=0.7.0",
    "scipy>=1.15.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "xformers>=0.0.33.post2",
]

